Organize the classes into suitable duke package
